wiserone::html

Function generate_html_file

pub fn generate_html_file(
    filename: &str,
    quote: &Quote,
) -> Result<(), Box<dyn Error>>

Creates an HTML file based on the provided quote.

Arguments

filename - The name of the file to be created (must be a simple filename, not a path).

quote - A reference to the quote to be used.

Returns

Returns Ok(()) if the file is successfully created, or an error otherwise.

Security

This function validates the filename to prevent directory traversal attacks. Files are always created in the designated output directory (./docs).